Resolve merge conflict from reorder
