+++ src/routes/index.ts
@@ -8,7 +8,6 @@ router.post("/conversations-post-event", controllers.conversationsPostEvent.post
 
 router.post("/inbound-call", controllers.inboundCall.post);
 
-router.post("/delete-session", controllers.session._delete);
 router.post("/sessions", controllers.session.post);
 
 export default router;
