--- src/utils/data.ts
@@ -1,1 +1,0 @@
-export function deleteMe() {}
